Create topview.py
The top view of a binary tree is the set of nodes visible when the tree is viewed from the top. Given a binary tree, print the top view of it. The output nodes can be printed in any order. A node x is there in the output if x is the topmost node at its horizontal distance. The horizontal distance of the left child of a node x is equal to a horizontal distance of x minus 1, and that of a right child is the horizontal distance of x plus 1. 
Approach: To solve the problem follow the below idea: 

The idea is to do something similar to vertical Order Traversal. Like vertical Order Traversal, we need to put nodes of the same horizontal distance together. We do a level order traversal so that the topmost node at a horizontal node is visited before any other node of the same horizontal distance below it. Hashing is used to check if a node at a given horizontal distance is seen or not. 

Follow the below steps to solve the problem:

Create a function to print the top view of the binary tree
If the root is equal to the null value then return from the function (Base case)
Create a queue of type Node* and a map of type <int, int> and a variable hd to calculate the horizontal distance
Set hd of the root equal to zero and push it into the queue
Run a while loop till the queue is not empty
set hd equal to root->hd
Check if this hd already exists in the map, If not then set this node->val in map[hd]
If root->left exists then set its hd = root->hd – 1 and push it into the queue
If root->right exists then set its hd = root->hd + 1 and push it into the queue
Pop the front element of the queue and set its value in the root
Print the answer using the map